🔧 코드라빗 리팩토링 제안 적용
- SPA 내비게이션 개선: <a> 태그를 Link 컴포넌트로 변경
- 컴포넌트 복잡도 개선: walk-full-test.tsx를 여러 컴포넌트로 분리
  - AuthenticationPanel: 인증 관련 UI 분리
  - WalkControlPanel: 산책 컨트롤 UI 분리
  - WalkRecordsList: 산책 기록 목록 UI 분리
  - useAuthentication: 인증 로직 커스텀 훅으로 분리
- useEffect 의존성 배열 개선: 타이머 관리 로직 최적화
- 액션 접근 방식 일관성 개선: useWalkActions 특정 액션만 반환
- API 키 검증 정규식 수정: 카카오맵 API 키 형식에 맞게 개선
- XSS 취약점 수정: dangerouslySetInnerHTML 제거하고 CSS 파일 분리

개선사항:
- 유지보수성 및 테스트 용이성 향상
- 타입 안전성 및 성능 최적화
- 코드 일관성 및 가독성 개선
- 보안 취약점 해결
