+++ src/pages/test/category-search.css
@@ -1,0 +1,63 @@
+/* 카카오맵 커스텀 오버레이 스타일 */
+.placeinfo_wrap {
+  position: absolute;
+  bottom: 28px;
+  left: -150px;
+  width: 300px;
+}
+
+.placeinfo {
+  position: relative;
+  width: 100%;
+  border-radius: 6px;
+  border: 1px solid #ccc;
+  border-bottom: 2px solid #ddd;
+  padding-bottom: 10px;
+  background: #fff;
+}
+
+.placeinfo:nth-of-type(1) {
+  border: 0;
+  box-shadow: 0px 1px 2px #888;
+}
+
+.placeinfo a {
+  color: #2c3e50;
+  text-decoration: none;
+}
+
+.placeinfo a:hover,
+.placeinfo a:active {
+  color: #21f1bf;
+  text-decoration: underline;
+}
+
+.placeinfo .title {
+  display: block;
+  overflow: hidden;
+  margin: 14px 0 0 10px;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-size: 14px;
+  font-weight: 700;
+}
+
+.placeinfo .tel {
+  color: #0f7833;
+}
+
+.placeinfo .jibun {
+  color: #999;
+  font-size: 11px;
+  margin-top: 0;
+}
+
+.after {
+  content: '';
+  position: relative;
+  margin-left: -12px;
+  left: 50%;
+  width: 22px;
+  height: 12px;
+  background: url('https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/vertex_white.png');
+}
